Viscacha 0.8 Beta 3

+++ admin/bbcodes.php
@@ -9,7 +9,9 @@ if ($job == 'smileys_delete') {
 	   	$result = $db->query('SELECT * FROM '.$db->pre.'smileys WHERE id IN ('.implode(',', $deleteid).')',__LINE__,__FILE__);
 	   	while ($row = $db->fetch_assoc($result)) {
 	   		$row['replace'] = str_replace('{folder}', $config['smileypath'], $row['replace']);
-	   		$filesystem->unlink($row['replace']);
+	   		if(file_exists($row['replace'])) {
+	   			$filesystem->unlink($row['replace']);
+	   		}
 	   	}
 		$db->query('DELETE FROM '.$db->pre.'smileys WHERE id IN ('.implode(',', $deleteid).')',__LINE__,__FILE__);
 		$anz = $db->affected_rows();
@@ -91,6 +93,14 @@ elseif ($job == 'smileys_import') {
   <td class="mbox"><input type="file" name="upload" size="40" /></td></tr>
   <tr><td class="mbox"><em>or</em> select a file from the server:<br /><span class="stext">Path starting from the Viscacha-root-directory: <?php echo $config['fpath']; ?></span></td>
   <td class="mbox"><input type="text" name="server" size="50" /></td></tr>
+  <tr><td class="mbox">Format of imported Smileypack:</td>
+  <td class="mbox"><select name="format">
+  <option value="viscacha_ini" selected="selected">Viscacha (Standard)</option>
+  <option value="phpbb2">phpBB 2</option>
+  <option value="none">No format</option>
+  </select></td></tr>
+  <tr><td class="mbox">Delete all existing smileys before import:</td>
+  <td class="mbox"><input type="checkbox" name="truncate" value="1" /></td></tr>
   <tr><td class="mbox">Delete file after import:</td>
   <td class="mbox"><input type="checkbox" name="delete" value="1" checked="checked" /></td></tr>
   <tr><td class="ubox" colspan="2" align="center"><input accesskey="s" type="submit" value="Import" /></td></tr>
@@ -102,34 +112,38 @@ elseif ($job == 'smileys_import') {
 elseif ($job == 'smileys_import2') {
 	$server = $gpc->get('server', none);
 	$del = $gpc->get('delete', int);
+	$format = $gpc->get('format', none);
+	$truncate = $gpc->get('truncate', int);
 	$inserterrors = array();
 	
 	if (!empty($_FILES['upload']['name'])) {
 		$filesize = 2048*1024;
-		$filetypes = array('.zip');
-		$dir = realpath('temp/');
+		$filetypes = array('zip');
+		$dir = realpath('temp').DIRECTORY_SEPARATOR;
 	
 		$insertuploads = array();
 		require("classes/class.upload.php");
 		 
 		$my_uploader = new uploader();
 		$my_uploader->max_filesize($filesize);
-		if ($my_uploader->upload('upload', $filetypes)) {
-			$my_uploader->save_file($dir, 2);
-			if ($my_uploader->return_error()) {
-				array_push($inserterrors,$my_uploader->return_error());
+		$my_uploader->file_types($filetypes);
+		$my_uploader->set_path($dir);
+		if ($my_uploader->upload('upload')) {
+			$my_uploader->save_file();
+			if ($my_uploader->upload_failed()) {
+				array_push($inserterrors,$my_uploader->get_error());
 			}
 		}
 		else {
-			array_push($inserterrors,$my_uploader->return_error());
+			array_push($inserterrors,$my_uploader->get_error());
 		}
-		$file = $dir.'/'.$my_uploader->file['name'];
+		$file = $dir.DIRECTORY_SEPARATOR.$my_uploader->fileinfo('filename');
 		if (!file_exists($file)) {
 			$inserterrors[] = 'File ('.$file.') does not exist.';
 		}
 	}
 	elseif (file_exists($server)) {
-		$ext = get_extension($server, true);
+		$ext = get_extension($server);
 		if ($ext == 'zip') {
 			$file = $server;
 		}
@@ -147,6 +161,7 @@ elseif ($job == 'smileys_import2') {
 	
 	$tempdir = 'temp/'.md5(microtime()).'/';
 	
+	// Exract
 	require_once('classes/class.zip.php');
 	$archive = new PclZip($file);
 	$failure = $archive->extract($tempdir);
@@ -159,12 +174,68 @@ elseif ($job == 'smileys_import2') {
 		error('admin.php?action=bbcodes&job=smileys_import', 'ZIP-archive could not be read or the folder is empty.');
 	}
 
-	$codes = array();
-	$result = $db->query('SELECT search FROM '.$db->pre.'smileys');
-	while ($row = $db->fetch_assoc($result)) {
-		$codes[] = $row['search'];
+	// Parse format
+	switch ($format) {
+		case 'phpbb2':
+			$package = array();
+			$d = dir($tempdir);
+			while (false !== ($entry = $d->read())) {
+				if (get_extension($entry) == 'pak') {
+					$lines = file($tempdir.$entry);
+					break;
+				}
+			}
+			$d->close();
+			$lines = array_map('trim', $lines);
+			foreach($lines as $line) {
+				$new_pack = array();
+				list($new_pack['replace'], $new_pack['desc'], $new_pack['search']) = explode('=+:', $line, 3);
+				$new_pack['replace'] = '{folder}/'.$new_pack['replace'];
+				$package[] = $new_pack;
+			}
+		break;
+		case 'none':
+			$package = array();
+			$d = dir($tempdir);
+			$i = 0;
+			while (false !== ($entry = $d->read())) {
+				if (@getimagesize($tempdir.$entry) != false) {
+					$i++;
+					$package[] = array(
+						'search' => ':smiley'.$i.':',
+						'replace' => '{folder}/'.$entry,
+						'desc' => ''
+					);
+				}
+			}
+			$d->close();
+		break;
+		default: // viscacha_ini
+			$package = $myini->read($tempdir.'/smileys.ini');
+		break;
 	}
-	$package = $myini->read($tempdir.'/smileys.ini');
+	
+	// Delete old smileys
+	$codes = array();
+	if ($truncate == 1) {
+	   	$result = $db->query('SELECT * FROM '.$db->pre.'smileys',__LINE__,__FILE__);
+	   	while ($row = $db->fetch_assoc($result)) {
+	   		$row['replace'] = str_replace('{folder}', $config['smileypath'], $row['replace']);
+	   		if(file_exists($row['replace'])) {
+	   			$filesystem->unlink($row['replace']);
+	   		}
+	   	}
+		$db->query('TRUNCATE TABLE '.$db->pre.'smileys',__LINE__,__FILE__);
+	}
+	else {
+		// Get existing smiley codes from database
+		$result = $db->query('SELECT search FROM '.$db->pre.'smileys');
+		while ($row = $db->fetch_assoc($result)) {
+			$codes[] = strtolower($row['search']);
+		}
+	}
+
+	// Copy files and prepare for inserting smileys
 	$sqlinsert = array();
 	foreach ($package as $ini) {
 		if (strpos($ini['replace'], '{folder}') !== false) {
@@ -172,13 +243,19 @@ elseif ($job == 'smileys_import2') {
 			$ini['replace_new'] = str_replace('{folder}', $config['smileypath'], $ini['replace']);
 			$n = 0;
 			while(file_exists($ini['replace_new'])) {
-				$ext = get_extension($ini['replace_new']);
+				$ext = get_extension($ini['replace_new'], true);
 				$length = strlen($ext);
 				$base = substr($ini['replace_new'], $length*(-1), $length);
 				$n++;
 				$base .= '_'.$n;
 				$ini['replace_new'] = $base.$ext;
 			}
+			$n = 0;
+			while(in_array($ini['search'], $codes)) {
+				$n++;
+				$ini['search'] = ':smiley'.$n.':';
+			}
+			$codes[] = $ini['search'];
 			$filesystem->copy($ini['replace_temp'], $ini['replace_new']);
 		}
 		$sqlinsert[] = '("'.$gpc->save_str($ini['search']).'", "'.$gpc->save_str($ini['replace']).'", "'.$gpc->save_str($ini['desc']).'")';
@@ -192,6 +269,9 @@ elseif ($job == 'smileys_import2') {
 	}
 	rmdirr($tempdir);
 	
+	$delobj = $scache->load('smileys');
+	$delobj->delete();
+		
 	ok('admin.php?action=bbcodes&job=smileys', $anz.' Smileys successfully imported.');
 }
 elseif ($job == 'smileys_export') {
@@ -254,7 +334,7 @@ elseif ($job == 'smileys') {
 <form name="form" method="post" action="admin.php?action=bbcodes">
  <table class="border">
   <tr> 
-   <td class="obox" colspan="6"><span style="float: right;">[<a href="admin.php?action=bbcodes&amp;job=smileys_import">Import Smileypack</a>]</span>Manage Smileys</td>
+   <td class="obox" colspan="6"><span style="float: right;">[<a href="admin.php?action=bbcodes&amp;job=smileys_import">Import Smileypack</a>]</span>Manage Smileys (<?php echo $db->num_rows($result); ?> Smileys)</td>
   </tr>
   <tr class="ubox">
    <td width="5%">Choose<br /><span class="stext"><input type="checkbox" onclick="check_all('id[]');" name="all" value="1" /> All</span></td>
@@ -357,26 +437,24 @@ elseif ($job == 'smileys_add') {
 	 
 	    $my_uploader = new uploader();
 		$my_uploader->max_filesize($filesize);
+		$my_uploader->file_types($filetypes);
+		$my_uploader->set_path($dir);
 		if (isset($imgwidth) && isset($imgheight)) {
 			$my_uploader->max_image_size($imgwidth, $imgheight);
 		}
-		if ($my_uploader->upload('upload_'.$i, $filetypes)) {
-			$my_uploader->save_file($dir, 2);
-			if ($my_uploader->return_error()) {
-				$error[] = $my_uploader->return_error();
-			}
-			else {
-				$has_upload = $gpc->save_str($my_uploader->fileinfo('name'));
+		if ($my_uploader->upload('upload_'.$i)) {
+			if ($my_uploader->save_file()) {
+				$has_upload = $gpc->save_str($my_uploader->fileinfo('filename'));
 			}
 		}
-		else {
-			$error[] = $my_uploader->return_error();
+		if ($my_uploader->upload_failed()) {
+			$error[] = $my_uploader->get_error();
 		}
 	}
 	if (strlen($gpc->get('code', str)) < 2) {
 		$error[] = 'Code is too short';
 	}
-	if (!$has_upload && strlen($img) < 5) {
+	if (empty($has_upload) || strlen($img) < 5) {
 		$error[] = 'Path of image is too short';
 	}
 	if (strlen($gpc->get('show', int)) != 1 && $gpc->get('show', int) != 0) {
@@ -603,8 +681,8 @@ elseif ($job == 'codefiles') {
 	$clang = array();
 	$d = dir("classes/geshi");
 	while (false !== ($entry = $d->read())) {
-		if (get_extension($entry,TRUE) == 'php' && !is_dir("classes/geshi/".$entry)) {
-			@include_once("classes/geshi/".$entry);
+		if (get_extension($entry) == 'php' && !is_dir("classes/geshi/".$entry)) {
+			include_once("classes/geshi/".$entry);
 			$short = str_replace('.php','',$entry);
 			$clang[$short]['file'] = $entry;
 			$clang[$short]['name'] = $language_data['LANG_NAME'];
@@ -665,6 +743,153 @@ elseif ($job == 'del_codefiles') {
 	$delobj = $scache->load('syntax-highlight');
 	$delobj->delete();
     ok('admin.php?action=bbcodes&job=codefiles', 'Files successfully deleted!');
+}
+elseif ($job == 'custombb_export') {
+	$id = $gpc->get('id', int);
+	
+	$result = $db->query("
+	SELECT bbcodetag, bbcodereplacement, bbcodeexample, bbcodeexplanation, twoparams, title, buttonimage 
+	FROM {$db->pre}bbcode 
+	WHERE id = '{$id}' 
+	LIMIT 1
+	", __LINE__, __FILE__);
+	$data = $db->fetch_assoc($result);
+	$data['button'] = null;
+	
+	if (!empty($data['buttonimage']) && (file_exists($data['buttonimage']) || preg_match('/^(http:\/\/|www.)([\wäöüÄÖÜ@\-_\.]+)\:?([0-9]*)\/(.*)$/', $data['buttonimage'])) ) {
+		if (preg_match('/^(http:\/\/|www.)([\wäöüÄÖÜ@\-_\.]+)\:?([0-9]*)\/(.*)$/', $data['buttonimage'])) {
+			$button = get_remote($data['buttonimage']);
+		}
+		else {
+			$button = file_get_contents($data['buttonimage']);
+		}
+		if ($button == REMOTE_CLIENT_ERROR || $button == REMOTE_INVALID_URL) {
+			$data['buttonimage'] = '';
+		}
+		else {
+		    $ext = get_extension($data['buttonimage']);
+			if (!in_array($ext, $imagetype_extension)) {
+				$data['buttonimage'] = '';
+			}
+			else {
+				$data['button'] = base64_encode($button);
+			}
+		}
+	}
+	else {
+		$data['buttonimage'] = '';
+	}
+	
+	$content = serialize($data);
+	
+	viscacha_header('Content-Type: text/plain');
+	viscacha_header('Content-Length: '.strlen($content));
+	viscacha_header('Content-Disposition: attachment; filename="'.$data['bbcodetag'].'.bbc"');
+	
+	print($content);
+}
+elseif ($job == 'custombb_import') {
+	echo head();
+	?>
+<form name="form2" method="post" enctype="multipart/form-data" action="admin.php?action=bbcodes&job=custombb_import2">
+ <table class="border" cellpadding="4" cellspacing="0" border="0">
+  <tr><td class="obox" colspan="2">Import a new Design</td></tr>
+  <tr><td class="mbox"><em>Either</em> upload a file:<br /><span class="stext">Allowed file types: .bbc - Maximum file size: <?php echo formatFilesize(1024*250); ?></span></td>
+  <td class="mbox"><input type="file" name="upload" size="40" /></td></tr>
+  <tr><td class="mbox"><em>or</em> select a file from the server:<br /><span class="stext">Path starting from the Viscacha-root-directory: <?php echo $config['fpath']; ?></span></td>
+  <td class="mbox"><input type="text" name="server" size="50" /></td></tr>
+  <tr><td class="mbox">Delete file after import:</td>
+  <td class="mbox"><input type="checkbox" name="delete" value="1" checked="checked" /></td></tr>
+  <tr><td class="ubox" colspan="2" align="center"><input accesskey="s" type="submit" value="Import" /></td></tr>
+ </table>
+</form>
+	<?php
+	echo foot();
+}
+elseif ($job == 'custombb_import2') {
+
+	$dir = $gpc->get('dir', int);
+	$server = $gpc->get('server', none);
+	$del = $gpc->get('delete', int);
+	$inserterrors = array();
+	
+	if (!empty($_FILES['upload']['name'])) {
+		$filesize = ini_maxupload();
+		$filetypes = array('bbc');
+		$dir = 'temp/';
+	
+		$insertuploads = array();
+		require("classes/class.upload.php");
+		 
+		$my_uploader = new uploader();
+		$my_uploader->max_filesize(1024*250);
+		$my_uploader->file_types($filetypes);
+		$my_uploader->set_path($dir);
+		if ($my_uploader->upload('upload')) {
+			if ($my_uploader->save_file()) {
+				$file = $dir.$my_uploader->fileinfo('filename');
+				if (!file_exists($file)) {
+					$inserterrors[] = 'File ('.$file.') does not exist.';
+				}
+			}
+		}
+		if ($my_uploader->upload_failed()) {
+			array_push($inserterrors,$my_uploader->get_error());
+		}
+	}
+	elseif (file_exists($server)) {
+		$ext = get_extension($server);
+		if ($ext == 'bbc') {
+			$file = $server;
+		}
+		else {
+			$inserterrors[] = 'The selected file is no BBC-file.';
+		}
+	}
+	else {
+		$inserterrors[] = 'No valid file selected.';
+	}
+	echo head();
+	if (count($inserterrors) > 0) {
+		error('admin.php?action=bbcodes&job=custombb_import', $inserterrors);
+	}
+	
+	$content = file_get_contents($file);
+	extract(unserialize($content));
+
+	if (empty($bbcodetag) || empty($bbcodereplacement) || empty($bbcodeexample)) {
+		error('admin.php?action=bbcodes&job=custombb_import', 'File not valid!');
+	}
+
+	$result = $db->query("SELECT * FROM {$db->pre}bbcode WHERE bbcodetag = '{$bbcodetag}' AND twoparams = '{$twoparams}'", __LINE__, __FILE__);
+	if ($db->num_rows($result) > 0) {
+		error('admin.php?action=bbcodes&job=custombb_import', 'There is already a BB-Code named &quot;'.$bbcodetag.'&quot;. You may not create duplicate names.');
+	}
+	
+	if (empty($button)) {
+		$buttonimage = '';
+	}
+	else {
+		$name = basename($buttonimage);
+		$buttonimage = "images/{$name}";
+		if (!file_exists($buttonimage)) {
+			$filesystem->file_put_contents($buttonimage, base64_decode($button));
+		}
+	}
+	
+	$db->query("
+	INSERT INTO {$db->pre}bbcode (bbcodetag, bbcodereplacement, bbcodeexample, bbcodeexplanation, twoparams, title, buttonimage)
+	VALUES ('{$bbcodetag}','{$bbcodereplacement}','{$bbcodeexample}','{$bbcodeexplanation}','{$twoparams}','{$title}','{$buttonimage}')
+	", __LINE__, __FILE__);
+	
+	if ($del > 0) {
+		$filesystem->unlink($file);
+	}
+	
+	$delobj = $scache->load('custombb');
+	$delobj->delete();
+	
+	ok('admin.php?action=bbcodes&job=custombb', 'BB-Code ('.$title.') successfully imported!');
 }
 elseif ($job == 'custombb_add') {
 	echo head();
@@ -854,15 +1079,21 @@ elseif ($job == 'custombb_edit2') {
 elseif ($job == 'custombb_delete') {
 	echo head();
 	$id = $gpc->get('id', int);
+	$result = $db->query("SELECT buttonimage FROM {$db->pre}bbcode WHERE id = '{$id}' LIMIT 1", __LINE__, __FILE__);
+	$image = $db->fetch_assoc($result);
 	?>
 	<table class="border" border="0" cellspacing="0" cellpadding="4" align="center">
 	<tr><td class="obox">Delete Custom BB Code</td></tr>
 	<tr><td class="mbox">
 	<p align="center">Do you really want to delete this custom BB code?</p>
 	<p align="center">
-	<a href="admin.php?action=bbcodes&job=custombb_delete2&id=<?php echo $id; ?>"><img border="0" align="middle" alt="" src="admin/html/images/yes.gif"> Yes</a>
-	&nbsp&nbsp;&nbsp;&nbsp&nbsp;&nbsp;
-	<a href="javascript: history.back(-1);"><img border="0" align="middle" alt="" src="admin/html/images/no.gif"> No</a>
+	<?php if (@file_exists($image['buttonimage']) && !preg_match('/^(http:\/\/|www.)([\wäöüÄÖÜ@\-_\.]+)\:?([0-9]*)\/(.*)$/', $image['buttonimage'])) { ?>
+	<a href="admin.php?action=bbcodes&amp;job=custombb_delete2&amp;id=<?php echo $id; ?>&amp;img=1"><img border="0" align="absmiddle" alt="" src="admin/html/images/yes.gif"> Yes, inclusive image</a><br />
+	<a href="admin.php?action=bbcodes&amp;job=custombb_delete2&amp;id=<?php echo $id; ?>"><img border="0" align="absmiddle" alt="" src="admin/html/images/yes.gif"> Yes, but not the image</a><br />
+	<?php } else { ?>
+	<a href="admin.php?action=bbcodes&amp;job=custombb_delete2&amp;id=<?php echo $id; ?>"><img border="0" align="absmiddle" alt="" src="admin/html/images/yes.gif"> Yes</a><br />
+	<?php } ?>
+	<br /><a href="admin.php?action=bbcodes&amp;job=custombb"><img border="0" align="absmiddle" alt="" src="admin/html/images/no.gif"> No</a>
 	</p>
 	</td></tr>
 	</table>
@@ -872,7 +1103,15 @@ elseif ($job == 'custombb_delete') {
 elseif ($job == 'custombb_delete2'){
 	echo head();
 	$id = $gpc->get('id', int);
-	$db->query("DELETE FROM {$db->pre}bbcode WHERE id = ".$id, __LINE__, __FILE__);
+	$img = $gpc->get('img', int);
+	if ($img == 1) {
+		$result = $db->query("SELECT buttonimage FROM {$db->pre}bbcode WHERE id = '{$id}' LIMIT 1", __LINE__, __FILE__);
+		$image = $db->fetch_assoc($result);
+		if (@file_exists($image['buttonimage']) && !preg_match('/^(http:\/\/|www.)([\wäöüÄÖÜ@\-_\.]+)\:?([0-9]*)\/(.*)$/', $image['buttonimage'])) {
+			$filesystem->unlink($image['buttonimage']);
+		}
+	}
+	$db->query("DELETE FROM {$db->pre}bbcode WHERE id = '{$id}' LIMIT 1", __LINE__, __FILE__);
 	$delobj = $scache->load('custombb');
 	$delobj->delete();
 	ok('admin.php?action=bbcodes&job=custombb', 'Custom BB Code successfully deleted');
@@ -883,7 +1122,7 @@ elseif ($job == 'custombb') {
 	?>
 	<table align="center" class="border">
 	<tr>
-		<td class="obox" align="center" colspan="4"><span style="float: right;">[<a href="admin.php?action=bbcodes&job=custombb_add">Add new BB Code</a>]</span>Custom BB Code Manager</td>
+		<td class="obox" align="center" colspan="4"><span style="float: right;">[<a href="admin.php?action=bbcodes&job=custombb_add">Add new BB Code</a>] [<a href="admin.php?action=bbcodes&job=custombb_import">Import BB Code</a>]</span>Custom BB Code Manager</td>
 	</tr>
 	<tr>
 		<td class="ubox" width="30%">Title</td>
@@ -894,7 +1133,7 @@ elseif ($job == 'custombb') {
 	<?php
 	while ($bbcode = $db->fetch_assoc($result)) {
 		if (!empty($bbcode['buttonimage'])) {
-			$src = "<img style=\"background: buttonface; border:solid 1px highlight;\" src=\"{$bbcode['buttonimage']}\" alt=\"\" />";
+			$src = "<img style=\"background-color: buttonface; border:solid 1px highlight;\" src=\"{$bbcode['buttonimage']}\" alt=\"\" />";
 		}
 		else {
 			$src = '-';
@@ -902,9 +1141,13 @@ elseif ($job == 'custombb') {
 		?>
 		<tr>
 			<td class="mbox"><?php echo $bbcode['title']; ?></td>
-			<td class="mbox"><code><?php echo $bbcode['bbcodeexample']; ?></code></td>
+			<td class="mbox"><code>[<?php echo $bbcode['bbcodetag'].iif($bbcode['twoparams'], '={option}'); ?>]{param}[/<?php echo $bbcode['bbcodetag']; ?>]</code></td>
 			<td class="mbox" align="center"><?php echo $src; ?></td>
-			<td class="mbox">[<a href="admin.php?action=bbcodes&job=custombb_edit&id=<?php echo $bbcode['id']; ?>">Edit</a>] [<a href="admin.php?action=bbcodes&job=custombb_delete&id=<?php echo $bbcode['id']; ?>">Delete</a>]</td>
+			<td class="mbox">
+			[<a href="admin.php?action=bbcodes&job=custombb_edit&id=<?php echo $bbcode['id']; ?>">Edit</a>] 
+			[<a href="admin.php?action=bbcodes&job=custombb_export&id=<?php echo $bbcode['id']; ?>">Export</a>] 
+			[<a href="admin.php?action=bbcodes&job=custombb_delete&id=<?php echo $bbcode['id']; ?>">Delete</a>]
+			</td>
 		</tr>
 		<?
 	}
